tricodex-ibm-python-fastapi/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
astroid==3.3.8
attrs==25.1.0
beautifulsoup4==4.13.3
beeai-framework==0.1.1
Brotli==1.1.0
bson==0.5.10
cachetools==5.5.2
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
chevron==0.14.0
click==8.1.8
dill==0.3.9
distro==1.9.0
dnspython==1.16.0
dotenv==0.9.9
duckduckgo_search==7.4.4
et_xmlfile==2.0.0
fastapi==0.115.8
filelock==3.17.0
frozenlist==1.5.0
fsspec==2025.2.0
google-auth==2.38.0
google-genai==1.2.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.1
hyperframe==6.1.0
ibm-cos-sdk==2.13.6
ibm-cos-sdk-core==2.13.6
ibm-cos-sdk-s3transfer==2.13.6
ibm-granite-community-utils @ git+https://github.com/ibm-granite-community/utils.git@5d67648927240b208a164d2466f0dc77200450e5
ibm_watson_machine_learning==1.0.367
ibm_watsonx_ai==1.2.8
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.0.0
isort==6.0.0
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
json_repair==0.39.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.19
langchain-core==0.3.37
langchain-ibm==0.3.6
langchain-text-splitters==0.3.6
langsmith==0.3.9
litellm==1.61.13
lomond==0.3.3
lxml==5.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mcp==1.3.0
mdurl==0.1.2
motor==3.7.0
multidict==6.1.0
networkx==3.4.2
numpy==1.26.4
openai==1.64.0
openpyxl==3.1.5
orjson==3.10.15
packaging==24.2
pandas==2.1.4
platformdirs==4.3.6
pluggy==1.5.0
propcache==0.3.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic-settings==2.8.0
pydantic_core==2.27.2
Pygments==2.19.1
pylint==3.3.4
pymongo==4.11.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
pyventus==0.6.0
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.2
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.23.1
rsa==4.9
safetensors==0.5.2
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.6
SQLAlchemy==2.0.38
sse-starlette==2.2.1
starlette==0.45.3
tabulate==0.9.0
tenacity==9.0.0
tiktoken==0.9.0
tokenizers==0.21.0
tomlkit==0.13.2
tqdm==4.67.1
transformers==4.49.0
types-requests==2.32.0.20241016
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
websockets==14.2
wheel==0.45.1
wikipedia==1.4.0
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
